Implement meson build sytem.
This commit brings the meson build system branch up-to-date
with the current CMake build system. It should be possible
now to build and test flatcc using the Meson build system, like so:

$ meson setup build
$ meson compile -C build
$ meson test -C build

I have tested directly on Windows and Linux, as I have these systems
at hand. In addition, a new GitHub CI job has been added to exercise it
on other platforms (macOS).

Further to this, meson has been added as an option to the build scripts
such that users can specify meson as their underlying build system.